Fixed the draw detection
It used to be the case that when your last play is on an already won field the game doesn't run the draw detection. Not it's like it was supposed to be. When you play on an already won field it doesn't run the win detection but the draw detection regardless.