Changed the maximum field size and improved the AI
Fixed the maximum field size to 4 because 5 is firstly to hard for an ai to calculate and secondly 5 fields are too hard to display on small mobile devices. The AI now adopts the different sizes better because the iterations now get calculated more dynamically with the field size. Before it was impossible to play against the AI with an field size over 4 because it was trying to look up to 20 steps in the future at an very early point of the game.