Simplified the create a lobby screen
To improve the users experience and reduce confusion while creating a new lobby we hid everything that's not neccessary to create a lobby. All what's left is the name of the lobby, if it's a public or an invite only lobby and the invite player name. The other options are hidden under a show more button.